Add owners to todos

Approved-by: Christopher Schmidt <csc@it-minds.dk>

+++ src/Services/TodoServices/TodoService.cs
@@ -11,16 +11,19 @@ namespace todoProject.Services.TodoServices
     {
         private readonly ApplicationDbContext _context;
         private readonly IMapper _mapper;
+        private readonly IUserResolver _userResolver;
 
-        public TodoService(ApplicationDbContext context, IMapper mapper)
+        public TodoService(ApplicationDbContext context, IMapper mapper, IUserResolver userResolver)
         {
             _context = context;
             _mapper = mapper;
+            _userResolver = userResolver;
         }
 
         public async Task<TodoListDto> CreateTodoAsync(TodoListDto todo)
         {
             var todoToCreate = _mapper.Map<Todo>(todo);
+            todoToCreate.OwnerId = _userResolver.CurrentUserId;
 
             await _context.AddAsync(todoToCreate);
             await _context.SaveChangesAsync();
@@ -31,6 +34,7 @@ namespace todoProject.Services.TodoServices
         public async Task<TodoListDto[]> GetAllTodosAsync()
         {
             var todos = await _context.Todos
+                .Where(t => t.OwnerId == _userResolver.CurrentUserId)
                 .ToArrayAsync();
             
             return _mapper.Map<TodoListDto[]>(todos);
@@ -38,8 +42,11 @@ namespace todoProject.Services.TodoServices
 
         public async Task<TodoListDto> GetById(int id)
         {
+            var currentUser = await _userResolver.GetCurrentUserAsync();
             var todo = await _context.Todos
                 .FindAsync(id);
+            
+            if (todo.OwnerId != currentUser.Id) return null;
             
             return _mapper.Map<TodoListDto>(todo);
         }
@@ -49,6 +56,7 @@ namespace todoProject.Services.TodoServices
             var todo = await _context.Todos.FindAsync(id);
 
             if (todo == null) return;
+            if (todo.OwnerId != _userResolver.CurrentUserId) return;
 
             _context.Todos.Remove(todo);
             await _context.SaveChangesAsync();
@@ -59,6 +67,7 @@ namespace todoProject.Services.TodoServices
             var originalTodo = await _context.Todos.FindAsync(updatedTodo.Id);
 
             if (originalTodo == null) return null;
+            if (originalTodo.Owner.Id != _userResolver.CurrentUserId) return null;
 
             originalTodo.Title = updatedTodo.Title;
             originalTodo.Done  = updatedTodo.Completed;
